requirements.txt
# SPDX-FileCopyrightText: 2026 Alexandre Gomes Gaigalas <alganet@gmail.com>
#
# SPDX-License-Identifier: ISC

apywire==0.2.0
starlette==0.52.1
uvicorn==0.40.0
PyYAML==6.0.3
apsw==3.51.2.0
